requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
mysqlclient==2.2.4
nanoid==2.0.0
packaging==24.1
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.7.0
